PR fix and emotion styled version downgrade due to incompatibility with storybook
