+++ package.json
@@ -30,8 +30,8 @@
         "@rollup/plugin-commonjs": "^15.1.0",
         "@rollup/plugin-node-resolve": "^9.0.0",
         "@rollup/plugin-url": "^6.0.0",
-        "@storybook/addon-essentials": "^6.1.1",
-        "@storybook/react": "^6.1.1",
+        "@storybook/addon-essentials": "^6.1.21",
+        "@storybook/react": "^6.1.21",
         "@testing-library/dom": "^7.26.3",
         "@testing-library/jest-dom": "^5.11.4",
         "@testing-library/react": "^11.0.4",
@@ -63,8 +63,9 @@
         "build"
     ],
     "dependencies": {
-        "@emotion/babel-preset-css-prop": "^11.0.0",
-        "@emotion/react": "^11.1.4",
+        "@emotion/babel-preset-css-prop": "10.0.27",
+        "@emotion/react": "11.1.5",
+        "@emotion/styled": "10.0.27",
         "@material-ui/core": "^4.11.2",
         "babel-eslint": "^10.1.0",
         "eslint": "^7.17.0",
